增加图片处理

+++ src/com/romantic/factor/base/entity/CoreInfo.java
@@ -1,11 +1,11 @@
 package com.romantic.factor.base.entity;
 
 public class CoreInfo {
-	private int id;
-	public int getId() {
+	private long id;
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 }
